Make some table utilities more general. Move them into common/lib.
