+++ backends/p4tools/modules/testgen/core/small_step/table_stepper.h
@@ -5,6 +5,7 @@
 #include <optional>
 #include <vector>
 
+#include "backends/p4tools/common/lib/table_utils.h"
 #include "ir/ir.h"
 #include "lib/cstring.h"
 
@@ -18,50 +19,14 @@ namespace P4Tools::P4Testgen {
 /// Implements small-step operational semantics for tables.
 class TableStepper {
  protected:
+    /// Reference to the calling expression stepper.
     ExprStepper *stepper;
 
+    /// The table for this particular stepper.
     const IR::P4Table *table;
 
-    /// KeyProperties define properties of table keys that are useful for execution.
-    struct KeyProperties {
-        /// The original key element
-        IR::KeyElement const *key;
-
-        /// The control plane name of the key.
-        cstring name;
-
-        /// The index of this key within the current table.
-        size_t index;
-
-        /// The match type of this key (exact, ternary, lpm, etc...)
-        cstring matchType;
-
-        /// Whether the key is tainted.
-        /// For matches such as LPM and ternary, we can still generate a match.
-        bool isTainted;
-
-        explicit KeyProperties(IR::KeyElement const *key, cstring name, size_t index,
-                               cstring matchType, bool isTainted)
-            : key(key), name(name), index(index), matchType(matchType), isTainted(isTainted) {}
-    };
-
     /// Basic table properties that are set when initializing the TableStepper.
-    struct TableProperties {
-        /// Control plane name of the table.
-        cstring tableName;
-
-        /// Whether key expressions of the table contain taint. This makes it difficult to match.
-        bool tableIsTainted = false;
-
-        /// Whether the table is constant and can not accept control plane entries.
-        bool tableIsImmutable = false;
-
-        /// Indicates whether the default action can be overridden.
-        bool defaultIsImmutable = false;
-
-        /// Ordered list of key fields with useful properties.
-        std::vector<KeyProperties> resolvedKeys;
-    } properties;
+    TableUtils::TableProperties properties;
 
  public:
     /* =========================================================================================
@@ -105,21 +70,6 @@ class TableStepper {
     /// @returns the result pointer of the friend class ExpressionStepper.
     ExprStepper::Result getResult();
 
-    /// The function compares the right and left expression.
-    /// Returns true if the values are equal or if the left value is more precise.
-    /// If the left value is constant entries and the right value
-    /// is Mask/Default then priority is given to constants.
-    /// For more detailed information see
-    /// https://github.com/p4lang/behavioral-model/blob/main/docs/simple_switch.md#longest-prefix-match-tables
-    /// @param lpmIndex indicates the index of the lpm entry in the list of entries.
-    /// It is used for comparison.
-    static bool compareLPMEntries(const IR::Entry *leftIn, const IR::Entry *rightIn,
-                                  size_t lpmIndex);
-
-    /// @returns whether the table is immutable, meaning control plane entries can not be added.
-    void checkTableIsImmutable();
-
-    /// Add the default action path to the expression stepper. If only hits if @param
     /// tableMissCondition is true.
     void addDefaultAction(std::optional<const IR::Expression *> tableMissCondition);
 
@@ -140,10 +90,9 @@ class TableStepper {
     /// match or does not match at all. The table stepper first checks these custom match types. If
     /// these do not match it steps through the default implementation. If it does not match either,
     /// a P4C_UNIMPLEMENTED is thrown.
-    virtual const IR::Expression *computeTargetMatchType(ExecutionState &nextState,
-                                                         const KeyProperties &keyProperties,
-                                                         TableMatchMap *matches,
-                                                         const IR::Expression *hitCondition);
+    virtual const IR::Expression *computeTargetMatchType(
+        ExecutionState &nextState, const TableUtils::KeyProperties &keyProperties,
+        TableMatchMap *matches, const IR::Expression *hitCondition);
 
     /// A helper function that computes whether a control-plane/table-key hits or not. This does not
     /// handle constant entries, it is specialized for control plane entries.
