+++ backends/p4tools/modules/testgen/targets/bmv2/table_stepper.h
@@ -57,7 +57,7 @@ class Bmv2V1ModelTableStepper : public TableStepper {
 
  protected:
     const IR::Expression *computeTargetMatchType(ExecutionState &nextState,
-                                                 const KeyProperties &keyProperties,
+                                                 const TableUtils::KeyProperties &keyProperties,
                                                  TableMatchMap *matches,
                                                  const IR::Expression *hitCondition) override;
