+++ backends/p4tools/modules/testgen/targets/pna/dpdk/table_stepper.cpp
@@ -7,8 +7,8 @@
 namespace P4Tools::P4Testgen::Pna {
 
 const IR::Expression *PnaDpdkTableStepper::computeTargetMatchType(
-    ExecutionState &nextState, const KeyProperties &keyProperties, TableMatchMap *matches,
-    const IR::Expression *hitCondition) {
+    ExecutionState &nextState, const TableUtils::KeyProperties &keyProperties,
+    TableMatchMap *matches, const IR::Expression *hitCondition) {
     // If the custom match type does not match, delete to the core match types.
     return SharedPnaTableStepper::computeTargetMatchType(nextState, keyProperties, matches,
                                                          hitCondition);
