+++ backends/p4tools/modules/testgen/targets/pna/dpdk/table_stepper.h
@@ -13,7 +13,7 @@ namespace P4Tools::P4Testgen::Pna {
 class PnaDpdkTableStepper : public SharedPnaTableStepper {
  protected:
     const IR::Expression *computeTargetMatchType(ExecutionState &nextState,
-                                                 const KeyProperties &keyProperties,
+                                                 const TableUtils::KeyProperties &keyProperties,
                                                  TableMatchMap *matches,
                                                  const IR::Expression *hitCondition) override;
